moving functions into separate files
