+++ packages/destination-actions/src/destinations/hubspot/customEvent/functions/hubspot-event-schema-functions.ts
@@ -1,5 +1,4 @@
-import { PayloadValidationError, IntegrationError, RetryableError, StatsContext } from '@segment/actions-core'
-import { SubscriptionMetadata } from '@segment/actions-core/destination-kit'
+import { PayloadValidationError, IntegrationError, RetryableError } from '@segment/actions-core'
 import {
   CreateEventDefinitionReq,
   CreatePropDefinitionReq,
@@ -10,72 +9,8 @@ import {
   CachableSchema,
   SchemaDiff,
   PropertyCreateResp
-} from './types'
-import { Client } from './client'
-import { LRUCache } from 'lru-cache'
-
-export const cache = new LRUCache<string, CachableSchema>({
-  max: 2000,
-  ttl: 1000 * 60 * 60
-})
-
-export function getSchemaFromCache(
-  name: string,
-  subscriptionMetadata?: SubscriptionMetadata,
-  statsContext?: StatsContext
-): CachableSchema | undefined {
-  if (!subscriptionMetadata || !subscriptionMetadata?.actionConfigId) {
-    statsContext?.statsClient?.incr('cache.get.error', 1, statsContext?.tags)
-    return undefined
-  }
-
-  const schema: CachableSchema | undefined = cache.get(`${subscriptionMetadata.actionConfigId}-${name}`) ?? undefined
-  return schema
-}
-
-export async function saveSchemaToCache(
-  schema: CachableSchema,
-  subscriptionMetadata?: SubscriptionMetadata,
-  statsContext?: StatsContext
-) {
-  if (!subscriptionMetadata || !subscriptionMetadata?.actionConfigId) {
-    statsContext?.statsClient?.incr('cache.save.error', 1, statsContext?.tags)
-    return
-  }
-
-  cache.set(`${subscriptionMetadata.actionConfigId}-${schema.name}`, schema)
-  statsContext?.statsClient?.incr('cache.save.success', 1, statsContext?.tags)
-}
-
-export function compareSchemas(schema1: Schema, schema2: CachableSchema | undefined): SchemaDiff {
-  if (schema2 === undefined) {
-    return { match: 'no_match', missingProperties: {} }
-  }
-
-  if (schema1.name !== schema2.name && schema1.name !== schema2.fullyQualifiedName) {
-    throw new PayloadValidationError("Hubspot.CustomEvent.compareSchemas: Schema names don't match")
-  }
-
-  const missingProperties: { [key: string]: SegmentProperty } = {}
-
-  for (const [key, prop1] of Object.entries(schema1.properties)) {
-    const prop2 = schema2.properties[key]
-    if (prop2 === undefined) {
-      missingProperties[key] = prop1
-      continue
-    }
-    if (prop1.stringFormat === prop2.stringFormat && prop1.type === prop2.type) {
-      continue
-    } else {
-      throw new PayloadValidationError("Hubspot.CustomEvent.compareSchemas: Schema property types don't match")
-    }
-  }
-
-  return {
-    match: Object.keys(missingProperties).length > 0 ? 'properties_missing' : 'full_match',
-    missingProperties
-  }
-}
+} from '../types'
+import { Client } from '../client'
 
 export async function getSchemaFromHubspot(client: Client, schema: Schema): Promise<CachableSchema | undefined> {
   const response = await client.getEventDefinition(schema.name)
